backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
motor==3.3.2
pymongo==4.6.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
httpx==0.26.0
tenacity==8.2.3
pytest==8.0.0
pytest-asyncio==0.23.5
psutil==5.9.8